Tidy CachedTreeHash trait
